mgr/dashboard: support inline edit for datatable
adding a new celltransformation that will transform the cell to a form
control when you click on the edit button which is inside a cell.

Added a new CellTemplate `editing` which if set to a cell, then it'll
add the edit button. Inorder for this to work, you will have to attach
your `formGroup` to `customTemplateConfig` like
```
customTemplateConfig: {
          formGroup: this.rbdForm
        }
```

You can add the validations as part of the formGroup and its subsequent
control.

Also using a `EditState` to keep track of the different cells I can edit
simultaneously in a single time.

Fixes: https://tracker.ceph.com/issues/72171
Signed-off-by: Nizamudeen A <nia@redhat.com>