mgr/dashboard: support inline edit for datatable
adding a new celltransformation that will transform the cell to a form
control when you click on the edit button which is inside a cell.

Added a new CellTemplate `editing` which if set to a cell, then it'll
add the edit button. You can also add validators to the control by using
the `customTemplateConfig` like
```
customTemplateConfig: {
          validators: [Validators.required]
        }
```

Also using a `EditState` to keep track of the different cells I can edit
simultaneously in a single time.

Fixes: https://tracker.ceph.com/issues/72171
Signed-off-by: Nizamudeen A <nia@redhat.com>