qa/suites/rados/thrash-old-clients: Add OSD warnings to ignore list
Add these to ignorelist:
- OSD_HOST_DOWN
- OSD_ROOT_DOWN

These warnings cause Teuthology tests to fail during OSD setup. They
are unrelated to RADOS testing and occur due to the nature of
thrashing tests, where OSDs temporarily mark themselves as down.

If an OSD is the only one in the cluster at the time, the cluster
may  incorrectly detect the host as down, even though other OSDs
are still starting up.

Fixes: https://tracker.ceph.com/issues/70972

Signed-off-by: Naveen Naidu <naveen.naidu@ibm.com>