radosgw-admin: Pass max_entries for bucket list
The changes in [1] did not take into account that
radosgw-admin code calls `RGWBucketAdminOp::info`
directly and passes a `RGWBucketAdminOpState`
struct where max_entries is not initialized so
we should not assume that it's zero.

This in turn broke the `bucket list --uid foo` and
`bucket stats --uid foo` commands as the output was
changed and thus not keeping backward compatibility.

This change makes sure that we populate max_entries
in `RGWBucketAdminOpState` if `--max-entries` argument
was specified otherwise we set it to zero to keep the
backward compatibility in the output format.

[1] https://github.com/ceph/ceph/pull/62777

Fixes: https://tracker.ceph.com/issues/72049
Signed-off-by: Tobias Urdin <tobias.urdin@binero.com>
(cherry picked from commit 3909c6554cdfcf1b05b5e32297b2e65e9c67af2b)