libcephfs_proxy: extend the protocol to support embedded permissions
This patch adds the changes to the protocol definition to support
sending the user credentials along with the request that requires it.

Using protocol version 1, instead of sending a pointer to a previously
allocated UserPerm structure, the caller will embed the uid, gid and the
list of additional groups in the request itself.

Signed-off-by: Xavi Hernandez <xhernandez@gmail.com>
(cherry picked from commit cde17574cf4c0e81e7eeab9874d1f319fdf4d89a)