qa/suites/krbd: use a standard fixed-1 cluster in unmap subsuite
A custom "fixed-1, but with the client on a separate node" cluster was
needed only for pre-single-major.yaml kernel which is no longer around.
This can be a single-node job now -- see commits 384503a92e88
("krbd/unmap: put client.0 on a separate remote") and 2885a47bea64
("qa/suites/krbd: drop pre-single-major test").

Signed-off-by: Ilya Dryomov <idryomov@gmail.com>