doc/man/8: Fix small errors and small improvements in mount.ceph.rst
Fix two cases of a stray space in the middle of syntax causing invalid
rendering in `man` output (introduced in #49630).

Fix a missing apostrophe in "its" and expand another "it's" to the same
"it is".

Harmonize usage of single and double backticks (for ceph commands
and parameters respectively) and quotation marks (for reference to
single characters in syntax).

Harmonize command parameter descriptions to always begin with capital
case and end in full stop.
Other minor capitalization fixes.

Fix "scale with" that supposedly should be "scales with".

Signed-off-by: Ville Ojamo <14869000+bluikko@users.noreply.github.com>