crimson/os/seastore/transaction: decrease Transaction::fresh_backref_extents when retiring pending backref extents
The reason that we count fresh backref extents is to prevent the
trim_alloc transaction from becoming too large.

However, at present, we only count the newly created backref extents,
ignoring that those extents may be retired during the trim_alloc
process.

One example in which this can be disastrous is that, when the clients
are issuing large scale deletes, the trim_alloc may remove large amounts
of backref mappings during which it may create a lot of backref extents
which are later removed as the removal continues. In this case, when the
counted fresh backref extents reaches the threshold there may only be a
relatively small a mount of live backref extents.

Signed-off-by: Xuehan Xu <xuxuehan@qianxin.com>