mgr/dashboard: load dashboard inventories asyncrhonously
decouple the minimal API into smaller APIs so that the dashboard can
asynchronously load the landing page by showing the things that we have
recieved initially. This will improve the initial loading time of the
dashboard page by atleast showing all the info except the osds and hosts
which is slower when tested in scale.

Signed-off-by: Nizamudeen A <nia@redhat.com>