osd/scrub: do not limit operator-initiated repairs
'auto-repair' scrubs are limited to a maximum of
'scrub_auto_repair_num_errors' damaged objects.
However, operator-initiated repairs should not be limited
by that number. Alas, a bug in a previous commit
(97de817) modified the code in such a way that it applied the
'scrub_auto_repair_num_errors' limit to all repairs,
including operator-initiated ones. This commit fixes that.

Fixes: https://tracker.ceph.com/issues/72438

Note: the fix is similar to 'Tentacle' & 'main' fixes
(PR#64860 & PR#64849), but - as the surrounding code
was changed, this is not a backport.

Signed-off-by: Ronen Friedman <rfriedma@redhat.com>