doc: add documentation about new health warning
This addition elaborates on the meaning behind
"Module failed to initialize" cluster errors.

The documentation advises users to try failing over
the mgr to reboot the module initialization process,
then if the error persists, file a bug report. I decided
to write it this way instead of providing more complex
debugging tips such as advising to disable some mgr modules
since every case will be different depending on which modules
failed to initialize.

In the bug report, developers can ask for the health detail
output to narrow down which module is causing a bottleneck,
and then ask the user to try disabling certain modules until
the mgr is able to fully initialize.

Signed-off-by: Laura Flores <lflores@ibm.com>