mgr/dashboard: fix cssStyleRule memory increasing
`getComputedStyle()` being called every change detection cycle increases
the cssStyleRule count in the memory. It returns a new object every time
the refresh cycle happen. In a long-lived tab, this cycles add up and
degrades the tab performance. So caching the style value so that
`[ngStyle]` doesn't think its new.

Signed-off-by: Nizamudeen A <nia@redhat.com>