erasure-code/consistency: Allow consistency checker to be able to deal with non-4k aligned buffers
A parity returned to the client will be the size of shard 0, which may not be 4k aligned in optimised erasure coding. A parity returned through the encode function will always be 4k aligned. This change makes it so we truncate the encoded data down to the size of the client write before comparing them together in the consistency check so we know we are comparing them like-for-like

Signed-off-by: Jon Bailey <jonathan.bailey1@ibm.com>