mgr, mon, osdc: pass complex parameters by rvalue reference
A non-trivial type is always passed by reference, even if the
parameter is declared as a value; in that case, the compiler must
allocate stack space and move the value there, and then really passes
a reference to this stack allocation.  This only adds overhead for no
reason.

A side effect of this API change is that all callers that accidently
created temporary copies are now revealed by throwing a compiler
error.  This commit fixes all callers and reduces a good amount of
useless overhead by wrapping those parameters in std::move() instead
of copying temporaries.

More temporaries are avoided by emplacing the std::string&& into the
std::vector.

Signed-off-by: Max Kellermann <max.kellermann@ionos.com>