cephfs.pyx: handle when UID/GID passed to chown() is -1
When chown() in libcephfs call receives -1 as the value of UID/GID, it
is implicitly converted from signed integer to unsigned integeer but
CephFS cython instead of doing that errors out during complilation.
Therefore make this conversion explicitly to allow users to pass -1
(which they do to indicate that UID/GID value shouldn't be changed) and
also for sake of uniformity and consistency between between CephFS
cython and libcephfs interface.

Fixes: https://tracker.ceph.com/issues/72547
Introduced-by: 65328aca31f2f5038cb602148120b9427370ed4f
Signed-off-by: Rishabh Dave <ridave@redhat.com>