.github/workflows/redmine-upkeep: bump limit
Now that this no longer hits the github API in general, it's safer to process
many more issues. This is generally good as it's not possible to construct a
specific search that will process issues that have had their corresponding PR
merged. (We do have merge triggers however.)  Additionally, process more issues
but only every 12 hours.

Signed-off-by: Patrick Donnelly <pdonnell@ibm.com>