crimson/.../fixed_kv_node: don't call copy_out if delta_buffer is empty
Cache::mark_transaction_conflicted calls get_delta(), which in turn
calls FixedKVNodeLayout::copy_out for lba nodes. If the mutation_pending
extent happens not to have any deltas, it'll fail in memcpy in
FixedKVNodeLayout::copy_out.

I think this is valid because a transaction may become conflicted
between when duplicate_for_write is called and when the actual mutation
is performed on the extent.

Fixes: https://tracker.ceph.com/issues/72579
Signed-off-by: Samuel Just <sjust@redhat.com>