cephadm: Use specified method for osd.default
When creating the first OSD in a cluster, the method requested was being
ignored - meaning an LVM OSD would be created in all cases. If a given cluster
couldn't support LVM, it could not be deployed. If we relay the method value
requested by the cephadm invocation, we can create OSDs as expected.

Fixes: https://tracker.ceph.com/issues/72381

Signed-off-by: Zack Cerza <zack@cerza.org>