mgr/cephadm: Version tracking with Cephadm
Instead of creating a separate class inheriting from MgrModule,
we move the functions and schema variables to cephadm.py, this is
used to leverage the MgrModule inheritance of cephadm.py and properly
integrate this feature as part of the cephadm module. Also added
version tracking to upgrade.py to add entries everytime a successful
upgrade occurs.

Signed-off-by: Timothy Q Nguyen <timqn22@gmail.com>